requirements.txt
# NeedScoop Dependencies

# ============================================================
# Core
# ============================================================
python-dateutil>=2.8.2
pyyaml>=6.0.1
click>=8.1.7
rich>=13.7.0  # Beautiful terminal output

# ============================================================
# Bluesky / AT Protocol
# ============================================================
atproto>=0.0.55

# ============================================================
# Reddit
# ============================================================
praw>=7.7.0

# ============================================================
# Vector Database
# ============================================================
chromadb>=0.4.22

# ============================================================
# Embeddings & ML (Local)
# ============================================================
sentence-transformers>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0

# ============================================================
# Clustering
# ============================================================
umap-learn>=0.5.5
hdbscan>=0.8.33

# ============================================================
# Data Processing
# ============================================================
pandas>=2.2.0
regex>=2023.12.25  # Better regex support

# ============================================================
# Visualization (Optional)
# ============================================================
matplotlib>=3.8.0
plotly>=5.18.0

# ============================================================
# Development
# ============================================================
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.1.0
ruff>=0.2.0
mypy>=1.8.0

# ============================================================
# Future: Web UI
# ============================================================
# streamlit>=1.31.0
# gradio>=4.16.0
